Back button, bounce to menu items
Great game here, really a neat project.
I added a couple things to but unfortunately I'm on a PC at the moment so this fork/pull request was typed in GitHub's editor (shouldn't be any typos).

Having a back button to get back to the main menu is pretty important for UX. I've added the code to the ClassicMenuScene but it should also be added to Zen and Stress as well. 

Also added a light bounce and delay to the following nodes as they enter the view from the bottom.